Fix TeamCity REST URL regex, relaxed requirements
The TeamCity REST URL was a tad broken. This PR fixes it.

The regex was also a little too restrictive for my needs in that it
would not allow me to specify more advanced build locator strings like
tags. That is fixed here in a way I hope to be in the spirit of the
original implementation.

Ideally it would be nice to just validate the `/app/rest/builds` part of
the URL. That should be enough to indicate the the user has entered
a valid TeamCity REST URI. Anyone with sufficiently advanced knowledge
of the TeamCity REST API might find anything more too restrictive.

Providing a helper string would be great, but a perhaps a link to the
following page could be presented to help people create their URI:

http://confluence.jetbrains.net/display/TW/REST+API+Plugin#RESTAPIPlugin-BuildLocator

I'm happy to tweak this PR as needed. Great tool!